fix: add event handler and apply to realtime client
